data/SimSwap/requirements.txt
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backcall
bash-kernel==0.9.3
beautifulsoup4
bleach==6.1.0
brotlipy==0.7.0
certifi==2020.12.5
cffi
chardet
charset-normalizer==3.3.2
coloredlogs==15.0.1
comm==0.2.0
conda==4.9.2
conda-build==3.21.4
conda-package-handling
contourpy==1.1.1
cryptography
cycler==0.12.1
debugpy==1.8.0
decorator
defusedxml==0.7.1
dnspython==2.1.0
easydict==1.11
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock
flatbuffers==23.5.26
fonttools==4.46.0
glob2
humanfriendly==10.0
idna
imageio==2.4.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
insightface==0.2.1
ipykernel==6.27.1
ipython==8.12.3
ipython-genutils
ipywidgets==8.1.1
jedi==0.17.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-archive==3.4.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.5.0
jupyter-events==0.9.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.1
jupyter-server==2.12.1
jupyter-server-terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy-loader==0.3
libarchive-c
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nbzip==0.1.0
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook-shim==0.2.3
numpy==1.24.4
olefile==0.46
onnx==1.15.0
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3
opencv-python==4.8.1.78
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso
pexpect
pickleshare
Pillow
pkginfo==1.7.0
pkgutil-resolve-name==1.3.10
platformdirs==4.1.0
pluggy==1.3.0
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil
ptyprocess
pure-eval==0.2.2
pycosat==0.6.3
pycparser
Pygments
pyOpenSSL
pyparsing==3.1.1
PySocks
pytest==7.4.3
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.7
pytz
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
ruamel-yaml==0.15.87
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
six
sniffio==1.3.0
soupsieve
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch==1.8.0
torchelastic==0.2.2
torchtext==0.9.0
torchvision==0.9.0
tornado==6.4
tqdm
traitlets==5.14.0
typing-extensions==4.9.0
tzdata==2023.3
urllib3
wcwidth
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
zipp==3.17.0
